refactoring the test of wallet creation to use page object pattern common used by appium.
