+++ app/src/androidTest/java/blockeq/com/stellarwallet/BasePage.kt
@@ -1,0 +1,5 @@
+package blockeq.com.stellarwallet
+
+abstract class BasePage {
+    internal abstract fun onPageLoaded(): BasePage
+}
