+++ app/src/androidTest/java/blockeq/com/stellarwallet/MnemonicType.kt
@@ -1,0 +1,6 @@
+package blockeq.com.stellarwallet
+
+enum class MnemonicType {
+    WORD_12,
+    WORD_24
+}
