+++ app/src/androidTest/java/blockeq/com/stellarwallet/WalletCreationTest.kt
@@ -1,0 +1,41 @@
+package blockeq.com.stellarwallet
+
+import android.support.test.rule.ActivityTestRule
+import android.support.test.runner.AndroidJUnit4
+import blockeq.com.stellarwallet.activities.LaunchActivity
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * IMPORTANT: make sure that the app and previous tests are not installed in the device.
+ * Run sh uninstallApk.sh first. This could be solved using orchestrator tests.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class WalletCreationTest {
+    private val pin = "1234"
+
+    @Rule
+    @JvmField
+    val activityTestRule = ActivityTestRule<LaunchActivity>(LaunchActivity::class.java)
+
+    @Test
+    fun testCreateWalletOption12Words() {
+        LaunchPage.onPageLoaded().createWallet(MnemonicType.WORD_12, pin)
+        WalletPage.onPageLoaded().pressSettings().clearWallet()
+        //restart
+        LaunchPage.onPageLoaded()
+    }
+
+    @Test
+    fun testCreateWalletOption24Words() {
+        LaunchPage.onPageLoaded().createWallet(MnemonicType.WORD_24, pin)
+        WalletPage.onPageLoaded().pressSettings().clearWallet()
+        //restart
+        LaunchPage.onPageLoaded()
+    }
+}
